requirements.txt
boto3
opencv-python
numpy
Pillow
moviepy
